Create Matrix.java
I wrote a Matrix class in Summer of 2014. This code is updated significantly from that. This version of the Matrix class supports complex numbers, for instance. I hope to add more methods. Note that the palu method (which finds the permuted LU factorization of the matrix) does NOT work correctly yet.